Restore Concurrent annotation, fix test failures.

+++ stack/core/src/test/java/org/apache/usergrid/persistence/PathQueryIT.java
@@ -78,10 +78,10 @@ public class PathQueryIT extends AbstractCoreIT {
         // query the users, ignoring page boundaries
         Results results = em.searchCollection( em.getApplicationRef(), "users", userQuery );
         PagingResultsIterator pri = new PagingResultsIterator( results );
-        int count = 2;
+        long count = 2;
         while ( pri.hasNext() ) {
             Entity e = ( Entity ) pri.next();
-            assertEquals( count++, (int)e.getProperty( "index" ) );
+            assertEquals( count++, e.getProperty( "index" ) );
         }
         assertEquals( count, expectedUserQuerySize + 2 );
